Move oidc and legacy sso code to separate `auth` package.
