+++ src/auth/legacy/sso/attemptCompleteLegacySsoLogin.test.ts
