+++ src/auth/legacy/sso/attemptCompleteLegacySsoLogin.ts
