+++ src/auth/legacy/sso/completeLegacySsoLogin.ts
