+++ src/auth/legacy/sso/startLegacySsoLoginFlow.ts
