+++ src/auth/oidc/registerOidcClient.ts
