+++ src/components/Welcome/WelcomePane.tsx
@@ -28,7 +28,8 @@ import {
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useSearchParams } from 'react-router';
-import { isValidServerName, startLoginFlow } from '../../lib';
+import { startLoginFlow } from '../../auth';
+import { isValidServerName } from '../../lib';
 import { Login } from '../Login/Login';
 import { WelcomeLogo } from './WelcomeLogo';
 import WelcomeSymbols from './WelcomeSymbols';
