+++ src/lib/testUtils/credentials.ts
@@ -16,8 +16,8 @@
  * along with NeoBoard Standalone. If not, see <https://www.gnu.org/licenses/>.
  */
 
+import { MatrixClientCredentials, OidcCredentials } from '../../auth';
 import { MatrixCredentials } from '../../state';
-import { MatrixClientCredentials, OidcCredentials } from '../oidc';
 
 export function mockMatrixClientCredentials(): MatrixClientCredentials {
   return {
